backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
google-generativeai==0.8.4
supabase==2.11.0
httpx==0.28.1
